Avoid test duplication and use RAII classes
